Separating out logging test and re-using test data
